requirements.txt
docx2txt==0.9
Flask==3.0.3
pillow==11.2.1
PyMuPDF==1.25.5
pytesseract==0.3.13
pytest==8.3.3
pytest-mock==3.14.0
python-magic==0.4.27
python-docx==1.1.2
RapidFuzz==3.13.0
ruff==0.11.7